[ADD] Added baseball.py file to freegames directory
Added 'baseball.py' file to free games directory which was
'test_baseball.py' in 'test' directory
